NvideaFProject (copy)/requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.0
langchain-huggingface==0.0.3
langserve==0.2.0
fastapi==0.111.0
uvicorn==0.30.0
gradio==4.36.0
chromadb==0.5.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
openai==1.30.0
pymupdf==1.24.0
pdfplumber==0.11.0
python-docx==1.1.0
python-dotenv==1.0.0
ragas==0.1.9
datasets==2.19.0
pytest==8.2.0
httpx==0.27.0
